DecoratedBrowser -> Browser, DecoratedPage -> Page
- Browser -> InternalBrowser, Page -> InternalPage
- setup launch()
- setup close()
