--- packages/browser/src/actions/index.ts
@@ -1,5 +1,0 @@
-export { clearElement } from "./clear";
-export { clickElement } from "./click";
-export { scrollElement } from "./scroll";
-export { selectElement } from "./select";
-export { typeElement } from "./type";
