--- packages/browser/src/browser/index.ts
@@ -1,1 +1,0 @@
-export * from "./Browser";
