+++ packages/browser/src/find/index.ts
@@ -1,7 +1,2 @@
 export * from "./find";
-export * from "./findCss";
-export * from "./findHtml";
 export * from "./FindOptionsBrowser";
-export * from "./findProperty";
-export * from "./findText";
-export * from "./hasText";
