+++ packages/browser/tests/actions/click.test.ts
@@ -1,12 +1,12 @@
 import { CONFIG } from "@qawolf/config";
-import { Browser } from "../../src/Browser";
-import { clickElement } from "../../src/actions";
+import { clickElement } from "../../src/actions/click";
+import { launch } from "../../src/browser/launch";
 import { hasText } from "../../src/find/hasText";
 
 describe("clickElement", () => {
   it("clicks on link", async () => {
-    const browser = await Browser.create({ url: CONFIG.testUrl });
-    const page = await browser.currentPage();
+    const browser = await launch({ url: CONFIG.testUrl });
+    const page = await browser.page();
 
     const element = await browser.find({ html: "<a>broken images</a>" });
     await clickElement(element);
@@ -18,8 +18,8 @@ describe("clickElement", () => {
   });
 
   it("clicks on icon in button", async () => {
-    const browser = await Browser.create({ url: `${CONFIG.testUrl}login` });
-    const page = await browser.currentPage();
+    const browser = await launch({ url: `${CONFIG.testUrl}login` });
+    const page = await browser.page();
 
     const hasInvalidUsernameText = await hasText(
       page,
