+++ packages/browser/tests/find/findText.test.ts
@@ -1,15 +1,14 @@
 import { CONFIG } from "@qawolf/config";
-import { Page } from "puppeteer";
-import { Browser } from "../../src/Browser";
-import { findText } from "../../src/find";
+import { Browser, launch, Page } from "../../src";
+import { findText } from "../../src/find/findText";
 import { getXpath } from "./utils";
 
 let browser: Browser;
 let page: Page;
 
 beforeAll(async () => {
-  browser = await Browser.create({ url: CONFIG.testUrl });
-  page = await browser.currentPage();
+  browser = await launch({ url: CONFIG.testUrl });
+  page = await browser.page();
 });
 
 afterAll(() => browser.close());
