+++ packages/browser/tests/find/hasText.test.ts
@@ -1,14 +1,13 @@
 import { CONFIG } from "@qawolf/config";
 import { Page } from "puppeteer";
-import { Browser } from "../../src/Browser";
-import { hasText } from "../../src/find";
+import { Browser, hasText, launch } from "../../src";
 
 let browser: Browser;
 let page: Page;
 
 beforeAll(async () => {
-  browser = await Browser.create({ url: `${CONFIG.testUrl}login` });
-  page = await browser.currentPage();
+  browser = await launch({ url: `${CONFIG.testUrl}login` });
+  page = await browser.page();
 });
 
 afterAll(() => browser.close());
