+++ packages/qawolf/src/index.ts
@@ -1,2 +1,20 @@
 #!/usr/bin/env node
+
+// run cli
 import "@qawolf/cli";
+
+// export public API
+export {
+  // actions
+  click,
+  Browser,
+  scroll,
+  select,
+  type,
+  // find
+  findProperty,
+  hasText,
+  // Browser
+  close,
+  launch
+} from "@qawolf/browser";
